+++ Models/FoodWasteReport.cs
@@ -10,9 +10,9 @@ namespace ReportEase.api.Models
         [BsonRepresentation(BsonType.ObjectId)]
         public string? Id { get; set; }
         public string? FoodItemId { get; set; }
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
         
-        public decimal Value { get; set; }
+        public double Value { get; set; }
         public string Description { get; set; }
         public string Department { get; set; }
         public string SubmittedBy { get; set; }
